Keep improving IoT2 example

+++ examples/IoT2/fr.inria.diverse.iot2.iot2/src-gen/fr/inria/diverse/iot2/iot2/aspects/Expression_CallFunctionAspect.java
@@ -62,7 +62,7 @@ public class Expression_CallFunctionAspect extends LuaExpressionAspect {
           final int max = Double.valueOf(_parseDouble_1).intValue();
           ThreadLocalRandom _current = ThreadLocalRandom.current();
           final int rand = _current.nextInt(min, max);
-          c.pushValue(Integer.valueOf(rand));
+          c.pushValue(Double.valueOf(((double) rand)));
           return;
         }
       }
